requirements.txt
# Selenium and Web Automation
selenium==4.15.2
webdriver-manager==4.0.1

# Data Handling
pandas==2.1.3
numpy==1.25.2

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Logging and Monitoring
loguru==0.7.2
colorlog==6.8.0

# HTTP Requests (backup)
requests==2.31.0
urllib3==2.1.0

# Development and Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
isort==5.12.0

# Type Hints
types-requests==2.31.0.10

# Utilities
click==8.1.7
tqdm==4.66.1
schedule==1.2.0